Implement script to download data
This script will allow download to resume, should download fail halfway.
This is useful as COCO dataset is quite large, and it's quite
inconvenient to re-download everything if the download fails halfway.

The fact that this script is also included in the makefile also allows
the downloaded data to be a dependency for other tasks, e.g. training
and deployment.

Finally, having a script to download the COCO dataset just results in
less hassle when having to setup this project elsewhere.